+++ src/context.rs
@@ -2,7 +2,7 @@ use crate::element::SvgElement;
 use crate::errors::{Result, SvgdxError};
 use crate::events::InputEvent;
 use crate::expression::eval_attr;
-use crate::position::BoundingBox;
+use crate::position::{BoundingBox, Size};
 use crate::types::{attr_split, extract_urlref, strp, AttrMap, ClassList, ElRef};
 use crate::TransformConfig;
 
@@ -149,6 +149,7 @@ impl Default for TransformerContext {
 pub trait ElementMap {
     fn get_element(&self, elref: &ElRef) -> Option<&SvgElement>;
     fn get_element_bbox(&self, el: &SvgElement) -> Result<Option<BoundingBox>>;
+    fn get_element_size(&self, el: &SvgElement) -> Result<Option<Size>>;
 }
 
 pub trait VariableMap {
@@ -164,6 +165,13 @@ impl ElementMap for TransformerContext {
             ElRef::Id(id) => self.elem_map.get(id),
             ElRef::Prev => self.prev_element.as_ref(),
         }
+    }
+
+    fn get_element_size(&self, el: &SvgElement) -> Result<Option<Size>> {
+        let target_el = el.get_target_element(self)?;
+        let el_size = target_el.size(self)?;
+
+        Ok(el_size)
     }
 
     fn get_element_bbox(&self, el: &SvgElement) -> Result<Option<BoundingBox>> {