+++ src/element.rs
@@ -10,7 +10,7 @@ use crate::events::{InputList, OutputEvent};
 use crate::expression::eval_attr;
 use crate::path::path_bbox;
 use crate::position::{
-    strp_length, BoundingBox, DirSpec, LocSpec, Position, ScalarSpec, TrblLength,
+    strp_length, BoundingBox, DirSpec, LocSpec, Position, ScalarSpec, Size, TrblLength,
 };
 use crate::text::process_text_attr;
 use crate::transform_attr::TransformAttr;
@@ -373,7 +373,23 @@ impl SvgElement {
             self.set_attr("d", &expand_relspec(&d, ctx));
         }
 
-        let p = Position::from(self as &SvgElement);
+        let mut p = Position::from(self as &SvgElement);
+        if self.name == "use" {
+            if let Some(href) = self.get_attr("href") {
+                let elref = href.parse()?;
+                let el = ctx
+                    .get_element(&elref)
+                    .ok_or_else(|| SvgdxError::ReferenceError(elref))?;
+                if let Some(sz) = ctx.get_element_size(el)? {
+                    p.update_size(&sz);
+                    if el.name == "circle" || el.name == "ellipse" {
+                        // The referenced element is defined by its center,
+                        // but use elements are defined by top-left pos.
+                        p.translate(sz.0 / 4., sz.1 / 4.);
+                    }
+                }
+            }
+        }
         p.set_position_attrs(self);
 
         Ok(())
@@ -699,7 +715,7 @@ impl SvgElement {
         Ok(element.clone())
     }
 
-    pub fn size(&self, ctx: &impl ElementMap) -> Result<Option<(f32, f32)>> {
+    pub fn size(&self, ctx: &impl ElementMap) -> Result<Option<Size>> {
         // NOTE: unlike bbox, this does not replace missing values with '0'.
         // Assumes any dw / dh have already been applied.
 
@@ -716,9 +732,15 @@ impl SvgElement {
         match self.name.as_str() {
             "use" | "reuse" => {
                 let target_el = self.get_target_element(ctx)?;
-                if let Ok(Some((w, h))) = target_el.size(ctx) {
-                    width = Some(w);
-                    height = Some(h);
+                // Take a _copy_ of the target element and evaluate attributes
+                // (should really only evaluate those which contribute to size...)
+                // This allows 'reuse' attributes which appear as vars within the
+                // target's context to determine size.
+                // let mut target_el = target_el.clone();
+                // target_el.eval_attributes(ctx)?;
+                if let Some(sz) = target_el.size(ctx)? {
+                    width = Some(sz.0);
+                    height = Some(sz.1);
                 }
             }
             "g" | "symbol" => {
@@ -762,7 +784,7 @@ impl SvgElement {
             _ => {}
         }
         if let (Some(width), Some(height)) = (width, height) {
-            Ok(Some((width, height)))
+            Ok(Some(Size(width, height)))
         } else {
             Ok(None)
         }
@@ -934,13 +956,6 @@ impl SvgElement {
                 "y" | "cy" | "y1" | "y2" => {
                     new_elem.set_attr(key, &fstr(strp(value)? + dy));
                 }
-                "points" => {
-                    let mut values = vec![];
-                    for (idx, part) in attr_split(value).enumerate() {
-                        values.push(fstr(strp(&part)? + if idx % 2 == 0 { dx } else { dy }));
-                    }
-                    new_elem.set_attr(key, &values.join(" "));
-                }
                 _ => (),
             }
         }
@@ -1091,7 +1106,7 @@ impl SvgElement {
         Ok((dx, dy))
     }
 
-    fn eval_rel_attributes(&mut self, ctx: &impl ElementMap) -> Result<()> {
+    pub fn eval_rel_attributes(&mut self, ctx: &impl ElementMap) -> Result<()> {
         for (key, value) in self.attrs.clone() {
             if self.is_size_attr(&key) {
                 let computed = self.eval_size_attr(&key, &value, ctx)?;
@@ -1181,7 +1196,7 @@ impl SvgElement {
                 let rel: DirSpec = reldir.parse()?;
                 // We won't have the full *position* of this element at this point, but hopefully
                 // we have enough to determine its size.
-                let (this_width, this_height) = self.size(ctx)?.unwrap_or((0., 0.));
+                let (this_width, this_height) = self.size(ctx)?.unwrap_or(Size(0., 0.)).as_wh();
                 let gap = if !remain.is_empty() {
                     let mut parts = attr_split(remain);
                     strp(&parts.next().unwrap_or("0".to_string()))?
@@ -1247,7 +1262,7 @@ impl SvgElement {
         }
     }
 
-    fn expand_compound_size(&mut self) {
+    pub fn expand_compound_size(&mut self) {
         if let Some(wh) = self.attrs.pop("wh") {
             // Split value into width and height
             let (w, h) = Self::split_compound_attr(&wh);
@@ -1305,7 +1320,7 @@ impl SvgElement {
     // xy="#o" -> x="#o", y="#o"
     // xy="#o 2" -> x="#o 2", y="#o 2"
     // xy="#o 2 4" -> x="#o 2", y="#o 4"
-    fn expand_compound_pos(&mut self) {
+    pub fn expand_compound_pos(&mut self) {
         // NOTE: must have already done any relative positioning (e.g. `xy="#abc|h"`)
         // before this point as xy is not considered a compound attribute in that case.
         if let Some(xy) = self.pop_attr("xy") {
@@ -1512,6 +1527,10 @@ mod tests {
 
         fn get_element_bbox(&self, el: &SvgElement) -> Result<Option<BoundingBox>> {
             el.bbox()
+        }
+
+        fn get_element_size(&self, el: &SvgElement) -> Result<Option<Size>> {
+            el.size(self)
         }
     }